requirements.txt
opencv-python==4.8.0.76
Pillow==10.0.0
face-recognition==1.3.0